+++ src/main/java/electrodynamics/common/tile/TileMineralGrinder.java
@@ -97,6 +97,6 @@ public class TileMineralGrinder extends GenericTile implements ITickableSound {
 	
 	@Override
 	public int getComparatorSignal() {
-		return (getNumActiveProcessors() / Math.max(1, getNumProcessors())) * 15;
+		return (int) (((double) getNumActiveProcessors() / (double) Math.max(1, getNumProcessors())) * 15.0);
 	}
 }
